delete names & locations
updated repository primary key type from Integer to Long for both WarehouseNameRepository and WarehouseLocationRepository

added findById function to handle invalid input
added delete function to remove WarehouseNames and WarehouseLocations using the findById function for chained valied input checking

+++ Back-End/src/main/java/com/mtlaguerre/inventory_management/repositories/WarehouseLocationRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.mtlaguerre.inventory_management.models.WarehouseLocation;
 
 @Repository
-public interface WarehouseLocationRepository extends JpaRepository<WarehouseLocation, Integer> {
+public interface WarehouseLocationRepository extends JpaRepository<WarehouseLocation, Long> {
 
 }
